applied the first template and the static file and created the first model

+++ shop/models.py
@@ -1,5 +1,7 @@
 from django.db import models
 
+
 class Item(models.Model):
     title = models.CharField(max_length=150)
-    price = models.IntegerField()
+    price = models.DecimalField(max_digits=6, decimal_places=2)
+
